backend/requirements.txt
fastapi
uvicorn
python-multipart
pydrive2
requests
pydantic
pdfplumber
pdf2image
pillow
pytesseract
paddleocr
python-docx
chromadb
pinecone-client
camelot-py
aiohttp
openai
google-generativeai
ollama
numpy
sentence-transformers
opencv-python
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
jinja2
